backend/requirements.txt
codecarbon==2.3.5
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.2
reportlab==4.2.2
certifi==2024.8.30
requests==2.32.3
python-dotenv==1.0.1
httpx==0.27.2
elevenlabs==1.2.0
textwrap3==0.9.2
